Fix for Swordapp library
- POST-ing to the SE-IRI is not supported by the official library, so we forked it to add support
- Refactored the code so that continued deposit is now routed through the container manager implementation
- Fixed the statement implementation
