+++ src/main/webapp/index.jsp
@@ -1,5 +1,10 @@
+<!DOCTYPE html>
 <html>
 <body>
-<h2>Hello World!</h2>
+<h1>Hello SWORDv2 World!</h1>
+<p>
+This is the EASY Deposit SWORDv2 service. For a description go to
+<a href="servicedocument">the Service Document</a>
+</p>
 </body>
 </html>
